Move things around & rename internal modules
Because it's always better to use full words and flat is better than
nested.
