+++ src/httpie_credential_store/__init__.py
@@ -1,6 +1,6 @@
 """HTTPie Credential Store."""
 
-from ._plug import CredentialStoreAuthPlugin, CredsAuthPlugin
+from ._plugin import CredentialStoreAuthPlugin, CredsAuthPlugin
 
 
 __all__ = ["CredentialStoreAuthPlugin", "CredsAuthPlugin"]
