+++ src/httpie_credential_store/_keychain.py
