+++ src/httpie_credential_store/_store.py
