+++ tests/test_plugin.py
@@ -6,13 +6,11 @@ import os
 import pathlib
 import re
 import sys
-import tempfile
 
 from urllib.request import parse_http_list, parse_keqv_list
 
 import pytest
 import responses
-import mock
 
 
 _is_windows = sys.platform == "win32"
@@ -42,24 +40,6 @@ class _RegExp(object):
 
     def __repr__(self):
         return self._regex.pattern
-
-
-@pytest.fixture(scope="session")
-def _httpie_config_dir():
-    """Set path to HTTPie configuration directory."""
-
-    # HTTPie can optionally read a path to configuration directory from
-    # environment variable. In order to avoid messing with user's local
-    # configuration, HTTPIE_CONFIG_DIR environment variable is patched to point
-    # to a temporary directory instead. But here's the thing, HTTPie is not ran
-    # in subprocess in these tests, and so the environment variable is read
-    # only once on first package import. That's why it must be set before
-    # HTTPie package is imported and that's why the very same value must be
-    # used for all tests (session scope). Otherwise, tests may fail because
-    # they will look for credentials file in different directory.
-    with tempfile.TemporaryDirectory() as tmpdir:
-        with mock.patch.dict(os.environ, {"HTTPIE_CONFIG_DIR": tmpdir}):
-            yield tmpdir
 
 
 @pytest.fixture(scope="function", autouse=True)
